# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards API
  description: Bank card management system with JWT authentication
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []

tags:
  - name: Auth
  - name: Admin Users
  - name: Admin Cards
  - name: Admin Block Requests
  - name: User Cards
  - name: Transfers

paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and obtain JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'

  /api/admin/users:
    post:
      tags: [Admin Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '200':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
    get:
      tags: [Admin Users]
      summary: List users
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Page of users

  /api/admin/users/{id}/enable:
    patch:
      tags: [Admin Users]
      summary: Enable or disable user
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEnableRequest'
      responses:
        '200':
          description: Updated user

  /api/admin/users/{id}/roles:
    patch:
      tags: [Admin Users]
      summary: Assign roles to user
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRolesRequest'
      responses:
        '200':
          description: Updated user

  /api/admin/cards:
    post:
      tags: [Admin Cards]
      summary: Create card for user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardCreateRequest'
      responses:
        '200':
          description: Created card (masked PAN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
    get:
      tags: [Admin Cards]
      summary: List all cards
      parameters:
        - name: ownerId
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/CardStatus'
        - name: search
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Page of cards

  /api/admin/cards/{id}:
    get:
      tags: [Admin Cards]
      summary: Get card by id
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
    delete:
      tags: [Admin Cards]
      summary: Delete card
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Deleted
    patch:
      tags: [Admin Cards]
      summary: Update card expiry fields
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardUpdateRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'

  /api/admin/cards/{id}/activate:
    patch:
      tags: [Admin Cards]
      summary: Activate card
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'

  /api/admin/cards/{id}/block:
    patch:
      tags: [Admin Cards]
      summary: Block card (admin)
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'

  /api/admin/block-requests:
    get:
      tags: [Admin Block Requests]
      summary: List pending block requests
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Page of block requests

  /api/admin/block-requests/{id}/approve:
    patch:
      tags: [Admin Block Requests]
      summary: Approve block request
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockRequestResponse'

  /api/admin/block-requests/{id}/reject:
    patch:
      tags: [Admin Block Requests]
      summary: Reject block request
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockRequestResponse'

  /api/cards:
    get:
      tags: [User Cards]
      summary: List own cards
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/CardStatus'
        - name: search
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Page of own cards

  /api/cards/{id}:
    get:
      tags: [User Cards]
      summary: Get own card
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'

  /api/cards/{id}/balance:
    get:
      tags: [User Cards]
      summary: Get card balance
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'

  /api/cards/{id}/block-request:
    post:
      tags: [User Cards]
      summary: Request card block (pending admin approval)
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockRequestResponse'

  /api/transfers:
    post:
      tags: [Transfers]
      summary: Transfer between own cards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    size:
      name: size
      in: query
      schema:
        type: integer
        default: 20

  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    JwtResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
        expiresIn:
          type: integer
          format: int64
    UserCreateRequest:
      type: object
      required: [username, password, roles]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        roles:
          type: array
          items:
            type: string
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        enabled:
          type: boolean
    UserEnableRequest:
      type: object
      properties:
        enabled:
          type: boolean
    UserRolesRequest:
      type: object
      properties:
        roles:
          type: array
          items:
            type: string
    CardCreateRequest:
      type: object
      required: [ownerId, pan, expiryMonth, expiryYear]
      properties:
        ownerId:
          type: integer
          format: int64
        pan:
          type: string
          pattern: '^\\d{16}$'
        expiryMonth:
          type: integer
        expiryYear:
          type: integer
        initialBalance:
          type: number
    CardUpdateRequest:
      type: object
      properties:
        expiryMonth:
          type: integer
          minimum: 1
          maximum: 12
        expiryYear:
          type: integer
          minimum: 2020
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maskedPan:
          type: string
          example: '**** **** **** 1234'
        ownerUsername:
          type: string
        expiryMonth:
          type: integer
        expiryYear:
          type: integer
        status:
          $ref: '#/components/schemas/CardStatus'
        balance:
          type: number
    CardStatus:
      type: string
      enum: [ACTIVE, BLOCKED, EXPIRED, BLOCK_REQUESTED]
    BalanceResponse:
      type: object
      properties:
        balance:
          type: number
    TransferRequest:
      type: object
      required: [fromCardId, toCardId, amount]
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number
          minimum: 0.01
    TransferResponse:
      type: object
      properties:
        fromCardId:
          type: integer
        toCardId:
          type: integer
        amount:
          type: number
        timestamp:
          type: string
          format: date-time
    BlockRequestResponse:
      type: object
      properties:
        id:
          type: integer
        cardId:
          type: integer
        ownerUsername:
          type: string
        maskedPan:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string